♻️ Move coordinator to a separate module
